feature: scrollable pages list
